requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==3.1.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
entrypoints==0.4
evdev==1.4.0
eyeD3==0.8.10
filelock==3.6.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
google-api-core==2.14.0
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.61.0
grpcio==1.30.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
meson==0.61.2
more-itertools==8.10.0
mpmath==0.0.0
mysql-connector-python==8.2.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nemo-emblems==5.8.0
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.11.0
notebook==6.4.8
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.0
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
tinycss2==1.1.1
tldextract==3.1.2
tornado==6.1
traitlets==5.1.1
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
Unidecode==1.3.3
uritemplate==3.0.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
